Modificacion en procesos logicos para integrar actualizacion futura de plugins
Se modificaron metodos que se volvern a utilizar en la comparacion entre registros, pero teniendo en cuenta que el registro local no se creara, sino que esta establecido en la carpet ups, y se cuido de no borrarlos para no inutilzar al updater. Se replico un macrometodo de logica comparativa pero teniendo en consideracion lo anterior, que el registro no se creara, si no que ya esta presente

+++ src/main/java/mo/updating/updaterRegisterComparison.java
@@ -23,7 +23,7 @@ public class updaterRegisterComparison {
      * @return true si hay diferencias en los archivos. Ya sea por tamaño, la no existencia de un archivo local comparado con el remoto o, alguna diferencia en la fecha de modificacion
      * @throws IOException
      */
-    public static boolean differencesInRegisters(Path localFile, Path remoteFile) throws IOException {
+    public static boolean differencesInRegisters(Path localFile, Path remoteFile, String mode) throws IOException {
 
         //CASO BASE: Algo sucedio con los registros de uso interno que, no estan disponibles o no existen desde procesos anteriores
         if( !Files.exists(localFile) || !Files.exists(remoteFile)) {
@@ -100,9 +100,13 @@ public class updaterRegisterComparison {
             //DEBUG: Descomentar si es necesario hacer una comparacion
             try {
                 // Elimina los registros que fueron sometidos a comparacion (argumentos de este metodo).
-                Files.deleteIfExists(localFile);
-                //Files.deleteIfExists(remoteFile);
+                if (mode.equals("mo")) {
+                    Files.deleteIfExists(localFile);
+                    System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                } else {
+                    //Files.deleteIfExists(remoteFile);
                 System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                }                
             } catch (IOException e) {
                 System.out.println("(updaterRegisterComparison.java) - Error al borrar los archivos de registro comparados" + e.getMessage());
             }
@@ -125,9 +129,13 @@ public class updaterRegisterComparison {
                 //DEBUG: Descomentar si es necesario hacer una comparacion
                 try {
                     // Elimina los registros que fueron sometidos a comparacion (argumentos de este metodo).
+                if (mode.equals("mo")) {
                     Files.deleteIfExists(localFile);
+                    System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                } else {
                     //Files.deleteIfExists(remoteFile);
                     System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                }  
                 } catch (IOException e) {
                     System.out.println("(updaterRegisterComparison.java) - Error al borrar los archivos de registro comparados" + e.getMessage());
                 }
@@ -155,9 +163,13 @@ public class updaterRegisterComparison {
                         //DEBUG: Descomentar si es necesario hacer una comparacion
                         try {
                             // Elimina los registros que fueron sometidos a comparacion (argumentos de este metodo).
-                            Files.deleteIfExists(localFile);
-                            //Files.deleteIfExists(remoteFile);
+                            if (mode.equals("mo")) {
+                                Files.deleteIfExists(localFile);
+                                System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                            } else {
+                                //Files.deleteIfExists(remoteFile);
                             System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+                            }  
                         } catch (IOException e) {
                             System.out.println("(updaterRegisterComparison.java) - Error al borrar los archivos de registro comparados" + e.getMessage());
                         }
@@ -185,9 +197,13 @@ public class updaterRegisterComparison {
         
         try {
             // Elimina los registros que fueron sometidos a comparacion (argumentos de este metodo).
-            Files.deleteIfExists(localFile);
-            //Files.deleteIfExists(remoteFile);
+            if (mode.equals("mo")) {
+                Files.deleteIfExists(localFile);
+                System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+            } else {
+                //Files.deleteIfExists(remoteFile);
             System.out.println("(updaterRegisterComparison.java) - Los archivos de registro comparados fueron borrados");
+            }  
         } catch (IOException e) {
             System.out.println("(updaterRegisterComparison.java) - Error al borrar los archivos de registro comparados" + e.getMessage());
         }
